+++ src/names.c
@@ -17,6 +17,12 @@ r_obj* vec_as_names(r_obj* names, const struct name_repair_opts* opts) {
   case NAME_REPAIR_unique: return vec_as_unique_names(names, opts->quiet);
   case NAME_REPAIR_universal: return vec_as_universal_names(names, opts->quiet);
   case NAME_REPAIR_check_unique: return check_unique_names(names, opts);
+  // At the time when unique_quiet and universal_quiet were added, no function
+  // that calls the C function vec_as_names() actually accepts these strings at
+  // the R level, because these functions enforce `quiet = false`.
+  // But we still have to handle every case for the enum.
+  case NAME_REPAIR_unique_quiet: return vec_as_unique_names(names, true);
+  case NAME_REPAIR_universal_quiet: return vec_as_universal_names(names, true);
   case NAME_REPAIR_custom: return vec_as_custom_names(names, opts);
   }
   r_stop_unreachable();
@@ -894,6 +900,12 @@ struct name_repair_opts new_name_repair_opts(r_obj* name_repair,
       opts.type = NAME_REPAIR_universal;
     } else if (c == strings_check_unique) {
       opts.type = NAME_REPAIR_check_unique;
+    } else if (c == strings_unique_quiet) {
+      opts.type = NAME_REPAIR_unique;
+      opts.quiet = true;
+    } else if (c == strings_universal_quiet) {
+      opts.type = NAME_REPAIR_universal;
+      opts.quiet = true;
     } else {
       struct repair_error_info info = new_repair_error_info(&opts);
       KEEP(info.shelter);
@@ -931,6 +943,8 @@ const char* name_repair_arg_as_c_string(enum name_repair_type type) {
   case NAME_REPAIR_unique: return "unique";
   case NAME_REPAIR_universal: return "universal";
   case NAME_REPAIR_check_unique: return "check_unique";
+  case NAME_REPAIR_unique_quiet: return "unique_quiet";
+  case NAME_REPAIR_universal_quiet: return "universal_quiet";
   case NAME_REPAIR_custom: return "custom";
   }
   r_stop_unreachable();